Add files via upload
This project analyzes daily sales data from a small grocery store for the month of May 2025. The goal is to understand which products and categories sell the most, which payment methods are used most often, and which days are the busiest. The data was cleaned using Excel, and the insights were visualized using Power BI through charts and interactive dashboards.